- Changed measure_openoffice_start_time.sh to use environment variables to pass file name to openoffice document. - Added calculating time difference.

+++ trunk/tools/openoffice/measure_openoffice_start_time.sh
@@ -1,14 +1,20 @@
 #!/bin/bash
 
-failure() {
-	echo "$@"
-	exit 1
-}
+THIS_SCRIPT_DIR=`dirname "$0"`
 
-RESULT_FILE=/tmp/start_time
-rm -f "$RESULT_FILE" || failure "Cannot remove $RESULT_FILE, error=$?"
+. "$THIS_SCRIPT_DIR"/../toolslib.sh
 
-echo "Start time: " `date --rfc-3339=ns`
+RESULT_FILE=$PWD/start_time.tmp
+#remove it as we check for its existence later
+if [ -e "$RESULT_FILE" ]; then
+	rm -f "$RESULT_FILE" || failure "Cannot remove $RESULT_FILE, error=$?"
+fi
+
+export OPENOFFICE_STARTUP_HOOK="$THIS_SCRIPT_DIR/openoffice_started.sh"
+export OPENOFFICE_STARTUP_HOOK_PARAMS="$RESULT_FILE"
+
+START_TIME=`get_time`
+echo "Start time: $START_TIME"
 
 oowriter ./recordStart.odt &
 
@@ -18,5 +24,11 @@ while true; do
 	fi
 	sleep 1
 done
+END_TIME=`cat "$RESULT_FILE"`
+
 echo "End time: " `cat "$RESULT_FILE"`
+echo "Difference:" `subtract_time $END_TIME $START_TIME`
+
+rm -f "$RESULT_FILE" || failure "Cannot remove $RESULT_FILE, error=$?"
+
 #killall soffice-bin
